Update C6cC reference data
I compared the three pion contractions at all momenta zero with
contractions that Liuming has performed a while ago. All diagrams match
except for the C6cC. Additionally we see that only that diagram suffered
from the strange `operator*` issues. Therefore we conclude that it was
wrong before and it is now fixed, therefore the reference data for that
diagram is updated now.