Multiply regardless of existence
The DilutedFactorsMap now has empty entries because they are
instantiated before. This means that they were skipped before and lead
to empty DilutedFactorMaps which lead to crashes in other parts of the
code. With the scheduling scheme we have there is no need to do it lazy
any more, we can just assume that they did not exist in the first place
and do the computations.